javascript: add an initial, incomplete example
This commit contains a semi-complete example of how to provide
client-side load balancing in the Javascript driver.
Currently the list of available nodes is hardcoded and
eventually it will need to be updated from alternator's /localnodes
API. Since everything in Javascript is virtual, the solution
is quite easy: the function responsible for looking up the address
is overridden with one that handles "dog.scylladb.com" in a special
way - by returning one of alternator's online nodes.
The code responsible for overriding lookups can be found in
`MoviesItemOps02.js`, which is the example file responsible for
GetItem requests.

Message-Id: <c35f1900b12b112a4f49875e5b13f1e7ec28daf6.1594209721.git.sarna@scylladb.com>