Add Go solution for Alternator load balancing
This is a fairly complete implementation of an Alternator load balancer
for almost-unmodified Go applications.

The only change required by the application is to change the way that
the AWS "session" is created once. Instead of the usual command:

   sess := session.Must(session.NewSessionWithOptions(session.Options{
               SharedConfigState: session.SharedConfigEnable,

The application should use something like:

    alternator_nodes := NewAlternatorNodes("http", 8000, []string {"127.0.0.1"})
    sess := alternator_nodes.session("dog.scylladb.com", "alternator", "secret_pass")

The meaning of these parameters is described in comments in the top of
alternator_lb.go.

When this session is used by the applications to send DynamoDB requests,
each request will be sent to a different Alternator node, in round-robin
fashion. Every second a background thread updates the list of Alternator
nodes.

BTW, node that in Go, unlike Python, this "session" object is thread-safe,
it can and will be used by an entire application and all its threads.

I think the implementation here is surprisingly elegant compared to other
languages. The AWS Go SDK provides various "hooks" into the request
process, and in particular allows us to write a "send handler", which
receives an HTTP request object and may modify it (including its URL)
before this request is run. So this is exactly what we do.

Luckily Go's HTTPRequest also allows us to easily modify the request's
Host header. Like in Java, they considered blindly using some random
"Host" header a security risk, but unlike Java they provided an
alternative mechanism to specify the host header (a separate "Host"
member in the request object), so we can easily override it, so this
implementation works well also with authentication.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>